chore: suite metadata types to common packages
