+++ suite-common/metadata-types/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "@suite-common/metadata-types",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "lint": "eslint '**/*.{ts,tsx,js}'",
+        "type-check": "tsc --build"
+    }
+}
