+++ suite-common/wallet-types/tsconfig.json
@@ -2,6 +2,9 @@
     "extends": "../../tsconfig.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
-        { "path": "../wallet-networks-config" }
+        { "path": "../wallet-networks-config" },
+        { "path": "../wallet-constants" },
+        { "path": "../intl-types" },
+        { "path": "../../packages/connect" }
     ]
 }
